+++ Cargo.toml
@@ -93,7 +93,6 @@ rolldown_error            = { path = "./crates/rolldown_error" }
 rolldown_fs               = { path = "./crates/rolldown_fs" }
 rolldown_oxc_utils        = { path = "./crates/rolldown_oxc_utils" }
 rolldown_plugin           = { path = "./crates/rolldown_plugin" }
-rolldown_rayon            = { path = "./crates/rolldown_rayon" }
 rolldown_resolver         = { path = "./crates/rolldown_resolver" }
 rolldown_rstr             = { path = "./crates/rolldown_rstr" }
 rolldown_sourcemap        = { path = "./crates/rolldown_sourcemap" }
