+++ crates/rolldown/src/utils/renamer.rs
@@ -2,8 +2,8 @@ use std::borrow::Cow;
 
 use oxc::semantic::ScopeId;
 use rolldown_common::{NormalModule, NormalModuleId, SymbolRef};
-use rolldown_rayon::{IntoParallelIterator, IntoParallelRefIterator, ParallelIterator};
 use rolldown_rstr::{Rstr, ToRstr};
+use rolldown_utils::rayon::{IntoParallelIterator, IntoParallelRefIterator, ParallelIterator};
 use rustc_hash::{FxHashMap, FxHashSet};
 
 use crate::types::{module_table::NormalModuleVec, symbols::Symbols};
