+++ crates/rolldown_utils/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name    = "rolldown_utils"
-version = "0.1.0"
+version = "0.0.1"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -11,3 +11,6 @@ repository.workspace = true
 
 [lints]
 workspace = true
+
+[target.'cfg(not(target_family = "wasm"))'.dependencies]
+rayon = { workspace = true }
