+++ crates/rolldown_utils/src/lib.rs
@@ -1,5 +1,6 @@
 // We keep some standalone utilities here
 
 mod bitset;
+pub mod rayon;
 
 pub use bitset::BitSet;
